Add listMarketBook API call

+++ betting_type/RunnerCatalog.cpp
@@ -5,7 +5,7 @@
 #include "RunnerCatalog.h"
 
 RunnerCatalog::RunnerCatalog(Json::Value json) {
-    selectionId = json["selectionId"].asInt();
+    selectionId = json["selectionId"].asInt64();
     runnerName = json["runnerName"].asString();
     handicap = json["handicap"].asDouble();
     sortPriority = json["sortPriority"].asInt();
